Snippets from current game
This is an example of my work;

MoveToLocation runs server side, handling all movement of players,
communicating with MoveToLocationHandler.cs which is client side. The
handler updates the player information client side as to the movement of
other players, etc via InGameController.cs. This handles all information
received for the player as their own information and other players
surrounding them and initiating updates to the server too.